Safari-611.1.10/LayoutTests/fast/css/variables/custom-property-computed-style-length-update-expected.txt

Check that CSSStyleDeclaration.length is accurate after custom property modifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS afterAddingComputedPropertyLength > initialLength is true
PASS afterAddingComputedPropertyLength === afterForcedLayoutLength is true
